Add textbox class and add feature to create multiple text boxes
